Move createJwt to utilities directory
